+++ apps/minifront/src/components/shared/non-native-fee-warning.tsx
@@ -8,9 +8,10 @@ import {
   getAmount,
   getAssetIdFromBalancesResponseOptional,
 } from '@penumbra-zone/getters/balances-response';
-import { viewClient } from '../../clients.ts';
+import { ViewService } from '@penumbra-zone/protobuf';
 import { GasPrices } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/component/fee/v1/fee_pb';
 import { getAssetId } from '@penumbra-zone/getters/metadata';
+import { penumbra } from '../../prax';
 
 const hasTokenBalance = ({
   source,
@@ -64,7 +65,7 @@ const useGasPrices = () => {
   const [prices, setPrices] = useState<GasPrices[]>([]);
 
   const fetchGasPrices = useCallback(async () => {
-    const res = await viewClient.gasPrices({});
+    const res = await penumbra.service(ViewService).gasPrices({});
     setPrices(res.altGasPrices);
   }, []);