+++ apps/minifront/src/fetchers/chain-id.ts
@@ -1,6 +1,7 @@
-import { viewClient } from '../clients';
+import { ViewService } from '@penumbra-zone/protobuf';
+import { penumbra } from '../prax';
 
 export const getChainId = async (): Promise<string | undefined> => {
-  const { parameters } = await viewClient.appParameters({});
+  const { parameters } = await penumbra.service(ViewService).appParameters({});
   return parameters?.chainId;
 };