+++ apps/minifront/src/fetchers/status.ts
@@ -1,10 +1,14 @@
-import { viewClient } from '../clients';
+import { ViewService } from '@penumbra-zone/protobuf';
+import { penumbra } from '../prax';
 
 const getInitialStatus = () =>
-  viewClient.status({}).then(status => ({
-    fullSyncHeight: status.fullSyncHeight,
-    latestKnownBlockHeight: status.catchingUp ? undefined : status.fullSyncHeight,
-  }));
+  penumbra
+    .service(ViewService)
+    .status({})
+    .then(status => ({
+      fullSyncHeight: status.fullSyncHeight,
+      latestKnownBlockHeight: status.catchingUp ? undefined : status.fullSyncHeight,
+    }));
 
 export async function* getStatusStream(): AsyncGenerator<{
   fullSyncHeight?: bigint;
@@ -15,7 +19,7 @@ export async function* getStatusStream(): AsyncGenerator<{
   // Therefore, we need to do a unary request to start us off.
   yield await getInitialStatus();
 
-  for await (const result of viewClient.statusStream({})) {
+  for await (const result of penumbra.service(ViewService).statusStream({})) {
     yield {
       fullSyncHeight: result.fullSyncHeight,
       latestKnownBlockHeight: result.latestKnownBlockHeight,