+++ apps/minifront/src/state/ibc-out.ts
@@ -6,7 +6,6 @@ import {
 import { BigNumber } from 'bignumber.js';
 import { ClientState } from '@buf/cosmos_ibc.bufbuild_es/ibc/lightclients/tendermint/v1/tendermint_pb.js';
 import { Height } from '@buf/cosmos_ibc.bufbuild_es/ibc/core/client/v1/client_pb.js';
-import { ibcChannelClient, ibcClient, ibcConnectionClient, viewClient } from '../clients';
 import {
   getAssetIdFromValueView,
   getDisplayDenomExponentFromValueView,
@@ -26,6 +25,13 @@ import { BLOCKS_PER_HOUR } from './constants';
 import { createZQuery, ZQueryState } from '@penumbra-zone/zquery';
 import { getChains } from '../fetchers/registry';
 import { bech32ChainIds } from './shared';
+import { penumbra } from '../prax';
+import {
+  IbcChannelService,
+  IbcClientService,
+  IbcConnectionService,
+  ViewService,
+} from '@penumbra-zone/protobuf';
 
 export const { chains, useChains } = createZQuery({
   name: 'chains',
@@ -136,7 +142,7 @@ const clientStateForChannel = async (channel?: Channel): Promise<ClientState> =>
     throw new Error('no connectionId in channel returned from ibcChannelClient request');
   }
 
-  const { connection } = await ibcConnectionClient.connection({
+  const { connection } = await penumbra.service(IbcConnectionService).connection({
     connectionId,
   });
   const clientId = connection?.clientId;
@@ -144,7 +150,9 @@ const clientStateForChannel = async (channel?: Channel): Promise<ClientState> =>
     throw new Error('no clientId ConnectionEnd returned from ibcConnectionClient request');
   }
 
-  const { clientState: anyClientState } = await ibcClient.clientState({ clientId: clientId });
+  const { clientState: anyClientState } = await penumbra
+    .service(IbcClientService)
+    .clientState({ clientId: clientId });
   if (!anyClientState) {
     throw new Error(`Could not get state for client id ${clientId}`);
   }
@@ -162,7 +170,7 @@ const clientStateForChannel = async (channel?: Channel): Promise<ClientState> =>
 const getTimeout = async (
   ibcChannelId: string,
 ): Promise<{ timeoutTime: bigint; timeoutHeight: Height }> => {
-  const { channel } = await ibcChannelClient.channel({
+  const { channel } = await penumbra.service(IbcChannelService).channel({
     portId: 'transfer',
     channelId: ibcChannelId,
   });
@@ -198,7 +206,9 @@ const getPlanRequest = async ({
   }
 
   const addressIndex = getAddressIndex(selection.accountAddress);
-  const { address: returnAddress } = await viewClient.ephemeralAddress({ addressIndex });
+  const { address: returnAddress } = await penumbra
+    .service(ViewService)
+    .ephemeralAddress({ addressIndex });
   if (!returnAddress) {
     throw new Error('Error with generating IBC deposit address');
   }